ESNSE: Scrapping Data Collection Economic Times
Issue #2

What's done?

A helper module is created in order to tackle the dynamically generated
content in economic times page. The module uses beautifulsoup and
selinium driver to execute javascript injection so that it can collect
the URLs alone as fast as possible.

The collected URLs are then passed to our scrapy project which scrapes
data and generates the json same like how it's implemented for
moneycontrol

Signed-off-by: Niteshkumar <nitesh156200@gmail.com>